separated concerns for SimpleLink
